+++ SuperBuild/External_Gadgetron.cmake
@@ -25,10 +25,11 @@
 set(proj Gadgetron)
 
 # Set dependency list
-set(${proj}_DEPENDENCIES "ACE;Boost;HDF5;ISMRMRD;FFTW3double;Armadillo;GTest")
+set(${proj}_DEPENDENCIES "ACE;Boost;HDF5;ISMRMRD;FFTW3double;Armadillo;GTest;range-v3;JSON;RocksDB;Date;mrd-storage-server")
 
 # Include dependent projects if any
 ExternalProject_Include_Dependencies(${proj} DEPENDS_VAR ${proj}_DEPENDENCIES)
+
 
 # Set external name (same as internal for now)
 set(externalProjName ${proj})
@@ -55,15 +56,32 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
 
   # BLAS
   find_package(BLAS)
-  if (APPLE AND NOT (CBLAS_LIBRARY AND CBLAS_INCLUDE_DIR))
-	# if the variables don't exist, let the user set them.
-	SET(CBLAS_LIBRARY "" CACHE FILEPATH "CBLAS library")
-	SET(CBLAS_INCLUDE_DIR "" CACHE PATH "CBLAS include directory")
-    message(FATAL_ERROR "Gadgetron needs CBLAS_LIBRARY and CBLAS_INCLUDE_DIR. If
-      these variables do not exist in your CMake, create them manually. CBLAS_LIBRARY
-      and CBLAS_INCLUDE_DIR should be FILEPATH and PATH, respectively, and live in
-      /usr/local/Cellar/openblas/ if installed with \"brew install openblas\".")
+
+  message(STATUS "CBLAS ${CBLAS_LIBRARY} ${CBLAS_INCLUDE_DIR}")
+  if (NOT (CBLAS_LIBRARY AND CBLAS_INCLUDE_DIR))
+    if (APPLE)
+      # if the variables don't exist, let the user set them.
+      SET(CBLAS_LIBRARY "" CACHE FILEPATH "CBLAS library")
+      SET(CBLAS_INCLUDE_DIR "" CACHE PATH "CBLAS include directory")
+      message(FATAL_ERROR "Gadgetron needs CBLAS_LIBRARY and CBLAS_INCLUDE_DIR. If
+        these variables do not exist in your CMake, create them manually. CBLAS_LIBRARY
+        and CBLAS_INCLUDE_DIR should be FILEPATH and PATH, respectively, and live in
+        /usr/local/Cellar/openblas/ if installed with \"brew install openblas\".")   
+    endif()
   endif()
+
+if (CBLAS_INCLUDE_DIR)
+  message(STATUS "Adding CBLAS_INCLUDE_DIR to Gadgetron_CMAKE_ARGS: ${CBLAS_INCLUDE_DIR}")
+  list (APPEND ${proj}_CMAKE_ARGS "-DCBLAS_INCLUDE_DIR:PATH=${CBLAS_INCLUDE_DIR}")
+else()
+  message(STATUS "CBLAS_INCLUDE_DIR will be found (probably) by Gadgetron")
+endif()
+if (CBLAS_LIBRARY)
+  message(STATUS "Adding CBLAS_LIBRARY to Gadgetron_CMAKE_ARGS: ${CBLAS_LIBRARY}")
+  list(APPEND ${proj}_CMAKE_ARGS "-DCBLAS_LIBRARY:FILEPATH=${CBLAS_LIBRARY}")
+else()
+    message(STATUS "CBLAS_LIBRARY will be found (probably) by Gadgetron")
+endif()
 
   #option(Gadgetron_BUILD_PYTHON_SUPPORT
   #  "Build Gadgetron Python gadgets (not required for SIRF)" OFF)
@@ -74,7 +92,13 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
     "Build Gadgetron MATLAB gadgets (not required for SIRF)" ${default_Gadgetron_BUILD_MATLAB_SUPPORT})
   option(Gadgetron_USE_MKL "Instruct Gadgetron to build linking to the MKL. The user must be able to install MKL on his own." OFF)
 
-  option(${proj}_USE_CUDA "Enable ${proj} CUDA (if cuda libraries are present)" ${USE_CUDA})
+  
+  if (USE_CUDA)
+    option(${proj}_USE_CUDA "Enable ${proj} CUDA" ${USE_CUDA})
+  else()
+    set (${proj}_USE_CUDA OFF)
+  endif()
+
   mark_as_advanced(${proj}_USE_CUDA)
 
   if (NOT DISABLE_OpenMP)
@@ -95,35 +119,33 @@ if(NOT ( DEFINED "USE_SYSTEM_${externalProjName}" AND "${USE_SYSTEM_${externalPr
 
   # Sets ${proj}_URL_MODIFIED and ${proj}_TAG_MODIFIED
   SetGitTagAndRepo("${proj}")
-
-  ExternalProject_Add(${proj}
-    ${${proj}_EP_ARGS}
-    ${${proj}_EP_ARGS_GIT}
-    ${${proj}_EP_ARGS_DIRS}
-
-    CMAKE_ARGS
-      -DBUILD_PYTHON_SUPPORT:BOOL=${Gadgetron_BUILD_PYTHON_SUPPORT}
-      -DBUILD_MATLAB_SUPPORT:BOOL=${Gadgetron_BUILD_MATLAB_SUPPORT}
+  set (${proj}_CMAKE_ARGS 
+      -DBUILD_PYTHON_SUPPORT:BOOL=${${proj}_BUILD_PYTHON_SUPPORT}
+      -DBUILD_MATLAB_SUPPORT:BOOL=${${proj}_BUILD_MATLAB_SUPPORT}
       -DCMAKE_PREFIX_PATH:PATH=${SUPERBUILD_INSTALL_DIR}
       -DCMAKE_LIBRARY_PATH:PATH=${SUPERBUILD_INSTALL_DIR}/lib
       -DCMAKE_INCLUDE_PATH:PATH=${SUPERBUILD_INSTALL_DIR}/include
-      -DCMAKE_INSTALL_PREFIX:PATH=${Gadgetron_INSTALL_DIR}
+      -DCMAKE_INSTALL_PREFIX:PATH=${${proj}_INSTALL_DIR}
       ${Boost_CMAKE_ARGS}
       ${PYTHONLIBS_CMAKE_ARGS}
       ${GTest_CMAKE_ARGS}
       ${HDF5_CMAKE_ARGS}
       ${FFTW3_CMAKE_ARGS}
       -DISMRMRD_DIR:PATH=${ISMRMRD_DIR}
-      -DUSE_MKL:BOOL=${Gadgetron_USE_MKL}
+      -DUSE_MKL:BOOL=${${proj}_USE_MKL}
       -DUSE_CUDA:BOOL=${${proj}_USE_CUDA}
-      -DCBLAS_INCLUDE_DIR:PATH=${CBLAS_INCLUDE_DIR}
-      -DCBLAS_LIBRARY:FILEPATH=${CBLAS_LIBRARY}
       -DUSE_OPENMP:BOOL=${${proj}_ENABLE_OPENMP}
-    DEPENDS
-        ${${proj}_DEPENDENCIES}
-    INSTALL_COMMAND ${CMAKE_COMMAND} --build . --target install 
-    COMMAND ${PYTHON_EXECUTABLE} ${CMAKE_SOURCE_DIR}/patches/Gadgetron_include-patch.py ${SUPERBUILD_INSTALL_DIR}/include/gadgetron/hoNFFT.h ${SUPERBUILD_INSTALL_DIR}/include/gadgetron/hoNFFT.h
-    COMMAND ${PYTHON_EXECUTABLE} ${CMAKE_SOURCE_DIR}/patches/copy_file_if_not_exists.py ${SUPERBUILD_INSTALL_DIR}/share/gadgetron/config/gadgetron.xml.example ${SUPERBUILD_INSTALL_DIR}/share/gadgetron/config/gadgetron.xml
+      -DBUILD_TESTING:BOOL=ON
+      )
+
+
+  ExternalProject_Add(${proj}
+    ${${proj}_EP_ARGS}
+    ${${proj}_EP_ARGS_GIT}
+    ${${proj}_EP_ARGS_DIRS}
+
+    CMAKE_ARGS ${${proj}_CMAKE_ARGS}
+    DEPENDS ${${proj}_DEPENDENCIES}
   )
 
     set(Gadgetron_ROOT        ${Gadgetron_SOURCE_DIR})